backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
tzdata==2025.2
